Current workaround for compiler crash.
For some reason nim cpp is having trouble with generics
surrounding VtValue, could be because of its weird type
erasure things, not sure just a wild guess. To get around
this, for now, typedesc can be used and hopefully, since
that information is all static the compiler can work out
the direct static calls that switch statement is really
doing.